e2e: move funs of framework/viperconfig to e2e
