--- test/e2e/framework/viperconfig/BUILD
@@ -1,36 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["viperconfig.go"],
-    importpath = "k8s.io/kubernetes/test/e2e/framework/viperconfig",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/github.com/spf13/viper:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["viperconfig_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//test/e2e/framework/config:go_default_library",
-        "//vendor/github.com/stretchr/testify/require:go_default_library",
-    ],
-)
